refactor: rename of card number input component to index
